Major bugfix in computeEnv
So far computeEnv (and integrate_x within computeEnv) were
using the state from the member variable "state". However,
the ODE solver does not change state. It maintains it's own
copy of the state until the step succeeds. Therefore, the
environment calculation in "derivs" does not see this
"updated" state from the solver, leading to errors. Hence,
explicitly pass the state (which derivs receives from the
ODE solver) to computeEnv thence to integrate_x.

This makes the computeEnv function slightly unelegant.
-- Maybe we should modify the ODE solver to instead update
the main state vector and maintain a backup copy in case the
step fails?