Commit to mark last birthFlux test --->
In this commit, the code to cross check the birthFLux calculation
within calcRates_XXX() functions is commented out, having asserted
that the calculations are now correct. The commented lines will be
removed in the next commit. For any hotfix related to this aspect
the comparison code can be obtained from this commit.