Begin modifications to allow extra state variables --->
Code design to allow specification of extra state variables is
quite non-trivial, due to performance considerations spanning all
three methods.

This branch begins to implement extra state, and will make
significant modifications to the code design. Might introduce
a separate species class and move many Solver methods to
Species.

This first commit implements a simple IteratorSet, which
provides named iterators to the begining of each variable
in the state array, and all iterators are advanced at once
with the ++ operator. The IteratorSet never exposes the
original iterators, so that the iterators can never be
independently incremented by the user.