Iterator set can now traverse custom layouts --->
Further improvements to IteratorSet. The constructor now takes
spacing and stride at inputs for each iterator, thus allowing
truely general and custom layouts to be traversed.

Removed map and do a linear search instead, since it it more
efficient for small arrays anyway.

Added a test for traversal speeds of different layouts,
which can be used to check cache-friendliness of the code.