Working env computation with single cohort--->
Using a dummy density of 1, environment computation can now be done
using solver provided integration functions. This seems to correctly
generate light profiles as the plant grows.

Modified interpolator to give clamped extrapolation. Maybe this should
be a control parameter for the interpolator.

added a logseq function in solver.tpp