Somewhat working CM --->
Plant model can now be simulated using CM method (x and u only).

However, outputs seem only approximately correct. It's quite
possible that this is due to log-transform of density used in
original model.

Saving the state here, and going to try the log-transform in next
commit in separate branch.