Multiples species in solver - working initialization
Solver now has a species vector and many grid related functions
are moved to species.

One issue here is that environment is common to all species
(there can be species specific perception of the environment
but the computeEnv function is responsible for managing the
species specific environments too). Hence computeEnv() can no
longer be a function within the Model class. One solution
which will be tried next is to supply a separate environment
object to the solver through another template parameter.

integrals will now also have to take a species index as input.