Working FMU solver --->
Environment needs to be supplied to g(), m(), f() functions,
because it is no longer part of Model. This is actually better,
because it makes the dependence of these functions on (x,t,E)
explicit.

Environment must also be set in the solver, added a Fn for that.

Minor bugfix in rates vector resizing.

fmu_calc_rates_dummy and fmu_u0in now working.

Next steps:
1) Maybe this is the right time to switch from x to Individual.
2) Solver and Species can store the model/env objects directly,
rather than storing pointers. In any case, these objects need not
linger outside the solver. getter functions can be added to get
pointers to these.