Working CM, but very slow --->
CM solver seems to be working, but is extremely slow!
Only after ode_eps is set to 1e-4 does it perform well.
Need to check this agaisnt the single-species version
or with the Plant model.

Profiling does not seem to work, but maybe thats a WSL
issue. Will have to check this on IIASA computer.

I will come back to these issues later.