1-cohort condition in size integrals now works --->
When there is only 1 cohort and when z < h(smallest cohort),
cohort at size xb should be included in the integral.
This was not working because the interpolator was crashing.

This was occuring for very high input seed rain values
because of wild oscillations in the interpolator at
very small sizes. Setting abs_tol to 1e-3 solves this issue.

as permanent solution, consider implementing constrained
cubic splines.