use iterators instead of vector-references in rate function --->
LSODA code operates on offsetted vectors, as it is derived from
underlying fortran code. Therefore, it passes &v[1] to the rate
funciton, but can be modified to pass an iterator.

Iterators are needed because we check the distance advanced
in calcRate_extra. Therefore, for now, I am making the rate
functions iterator-based, instead of double* based.

Except for really old solvers, modern ODE solvers should be able
to pass around iterators.