Working CM demos and error analysis --->
For CM to work in feedback mode, cohort removal has to be done
using both dx_cut and u_cut.

In RED model, removing densest cohort does not work: because
of log scale, dx is always smallest for the newest cohort, and
the (second) most recent cohort always gets removed.