[critical bugfix] in Hashtable, copy assignment operator was not returning --->
For the Plant demo,
In Hashtable class, copy assignment operator was missing a return *this
statement, which was leading to a segfault. Fixed.

Also noted here: After streamlining of the birthflux calcs in CM, the
seed rains in the Plant model changed slightly.

Before:   12.1881    19.8324    3.69066
Now:      12.1883    19.8322    3.6867

In the 'before' version, birthflux was calculated only at the insertion of
new cohort. 'Now' it is calculated within rate calc. This affects the
computation of the size integral, which uses boundary u. 'Now' is more
correct.