[generic_nd] plant demos working with feedback/IC; resolve EBT/IEBT issue with longterm artefact -->
In Plant, EBT and IEBT methods were predicting an artefactual high
seed rain for 1st species. This was because of a few very old
cohorts that had reached a cumulative mortality of Inf, and as a
result, their mort rate was getting set to 0. Thus, they kept
producing seeds. This is now fixed.

Gladly, the problem was in the Plant model implementation and not
in the EBT solvers.