[develop] hacky bugfix in implicit ODE stepper; move rk45ck::hmin to class member --->
In implicit methods, after stepping U, ODE stepper is used to step
accumulators. In that case, since hmin in RK45CK is set to 1e-6,
the odeStepper sometimes oversteps by 1e-6.

This causes subtle problems in the cohort insertion schedule, sometimes
necessitating substeps of 0.000001. For now, this is fixed in a
hacky way by explicitly setting the new time after the ode call.

Maybe a cleaner solution exists.